requirements.txt
fastapi
uvicorn
scikit-learn==1.2.2
numpy==1.26.4
pandas==2.2.3
openai
modal
joblib
pydantic
python-dotenv
pytest
requests
gunicorn  # Optional, for production deployments
